Add full mocks for general and gpio
Currently developing a test framework for the OV2640, which requires
STM32 HAL functions to be mocked. SPI and I2C is currently stubbed, but
will be expanded on later. A CMakeList framework for the build was also
created with the intention of adding cmocka into it for unit testing
later.

Signed-off-by: Adnan Soudki <asoudki@uwaterloo.ca>